Moved templates to router and general refactoring
