+++ .gitignore
@@ -1,2 +1,4 @@
 coverage/
 node_modules
+.idea/*
+*.iml
