+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "swagger-router",
-  "version": "0.1.2",
+  "version": "0.2.0",
   "description": "An efficient swagger 2 based router with support for multiple APIs. For use in RESTBase.",
   "main": "index.js",
   "scripts": {
@@ -24,15 +24,18 @@
     "url": "https://github.com/gwicke/swagger-router/issues"
   },
   "dependencies": {
-    "bluebird": "^2.7.1",
-    "es6-shim": "^0.22.1",
-    "js-yaml": "~3.2.2"
+    "bluebird": "2.8.2",
+    "core-js": "^1.1.4",
+    "js-yaml": "^3.4.2",
+    "tassembly": "^0.2.0",
+    "template-expression-compiler": "^0.1.0"
   },
   "devDependencies": {
-    "mocha": "x.x.x",
-    "mocha-jshint": "0.0.9",
-    "istanbul": "0.3.5",
-    "mocha-lcov-reporter": "0.0.1",
-    "coveralls": "2.11.2"
+    "mocha": "^2.2.5",
+    "mocha-jshint": "^2.2.3",
+    "istanbul": "^0.3.15",
+    "mocha-lcov-reporter": "^0.0.2",
+    "coveralls": "^2.11.2",
+    "mocha-jscs": "^1.2.0"
   }
 }
