+++ test/mocha.opts
@@ -1,0 +1,2 @@
+--recursive
+--no-timeouts
